Provide `with.user: __token__`
The gh-action-pypi-publish plugin docs seem to indicate that `with.user: __token__` should be provided when using an API Token. 

(I thought this was causing me 'Forbidden 403' issues, but it may have been a red herring. In my case, my API Token GitHub secret actually needed to be replaced to fix my 403 error.)